Add files via upload
This project focuses on the early detection of pests on the crops using the image processing techniques.The Training and Testing of dataset is done by the CNN classifier which trains the data. And each image undergoes various image processing operations to calculate the final density of pests on each crop.